refactor(seo): single source of truth for deposit rails
The rail list lived in two places — DEPOSIT_RAILS (exchanges.ts) and a
hand-copied RAIL_SLUGS Set in scripts/verify-content.ts, guarded only by
a 'keep in sync' comment. Removing avalanche meant editing both; the
lists had already drifted before.

Extract to src/data/seo/deposit-rails.ts — an import-free module both
the app (routes + sitemap via exchanges.ts) and the standalone content
gate (verify-content.ts) import. Add/remove a rail in one place now.

No behaviour change: same 9 rails, same classification. (Registering
faster-payments + spei as rails is a follow-up — it's a content-input
regeneration job, not a code add: the from-→via- flip breaks generated
RelatedLinks that must be fixed at the mono input layer and re-mirrored.)
