+++ src/data/seo/exchanges.ts
@@ -4,13 +4,14 @@
 //   - exchanges (binance, coinbase, …) — entity-backed in mono
 //   - rails (sepa, ach, arbitrum, …)   — pure content, no entity
 //
-// We split them by intersecting against the static rail list below. Display
+// We split them by intersecting against RAIL_SLUGS (see ./deposit-rails). Display
 // names + recommended network come from content frontmatter fields denormalized
 // at generation time (see mono/content/_system/templates/intents/deposit-from.md);
 // `recommended_network:` is the editorial pick for the lowest-friction path
 // from this exchange to Peanut. Absent values fall through to title-casing.
 
 import { listContentSlugs, readPageContent } from '@/lib/content'
+import { RAIL_SLUGS } from './deposit-rails'
 import { displayNameFromContent } from './utils'
 
 export interface Exchange {
@@ -23,32 +24,6 @@ interface DepositFrontmatter {
     recommended_network?: unknown
     published?: boolean
 }
-
-/** Crypto networks + fiat rails served at /deposit/via-{slug}. Hardcoded
- *  because rails have no entity data — they're purely a content-page concept.
- *
- *  TODO(reorg): two missing fiat keys need to be added during the next pass:
- *      'faster-payments': 'Faster Payments',  // UK — GBP via Bridge, live
- *      spei: 'SPEI Bank Transfer',            // Mexico — MXN via Bridge, live
- *  MDX content already exists at mono/content/deposit/{spei,faster-payments}/
- *  (pushed 2026-05-25). The pages 404 on the live site until the keys are
- *  registered here — generateStaticParams iterates Object.keys(DEPOSIT_RAILS).
- *  Both rails are wired end-to-end already: see src/utils/bridge.utils.ts
- *  getCurrencyConfig('MX' | 'GB', ...) — onramp + offramp via Bridge.
- *  Full context: mono/content/_system/ROADMAP.md (entry dated 2026-05-22). */
-export const DEPOSIT_RAILS: Record<string, string> = {
-    ach: 'ACH Bank Transfer',
-    sepa: 'SEPA Bank Transfer',
-    wire: 'Wire Transfer',
-    arbitrum: 'Arbitrum',
-    base: 'Base',
-    ethereum: 'Ethereum',
-    polygon: 'Polygon',
-    solana: 'Solana',
-    tron: 'Tron',
-}
-
-const RAIL_SLUGS = new Set(Object.keys(DEPOSIT_RAILS))
 
 function pickRecommendedNetwork(fm: DepositFrontmatter | undefined): string {
     const raw = fm?.recommended_network
